requirements.txt
# LLM — Groq (free, fast)
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Embeddings — local, free, no API key needed
sentence-transformers>=3.0.0

# Vector store
faiss-cpu>=1.7.4

# Sparse retrieval
rank-bm25>=0.2.2

# API & UI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.35.0

# Document loading
pypdf>=4.2.0
python-multipart>=0.0.9

# Utilities
pydantic>=2.7.0
httpx>=0.27.0
numpy>=1.26.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
